+++ CHANGELOG.md
@@ -3,8 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased - master]
+- Commands:
+    - Added support for DELETE command
+    - Added support for VERSION command
 - Swap to relative imports
-- Added support for VERSION command
 - Fixed client hanging after sending requests due to bad
   terminator chars